This commit includes several major changes to the validation within the project - both php and javascript.
1. Added errorDisplayOption form attribute which currently can be used
to toggle between the two supported options for displaying errors.
Option 0 is displaying errors above the appropriate element.  Option 1
is displaying errors in an unordered list at the top of the form.  In
non-mapping mode, the project will default to using option 0.  When the
map form attribute is used, option 1 will be the default.

2. Added renderAjaxErrorResponse function which is used to facilitate
php validation when the form's data is being submitted via ajax.  This
function builds and echos/returns a JSON string that is then
automatically parsed and applied to the form.

NOTE: Unfortunately, the example files are a mess right now from my
testing.  These will be corrected in my next commit.